ingestion/requirements.txt
pdfplumber>=0.11.4
pdf2image>=1.17.0
pytesseract>=0.3.10
pillow>=10.2.0
sentence-transformers>=3.0.1
tqdm>=4.66.4
numpy>=1.26.4
python-dotenv>=1.0.1
supabase>=2.3.3
typer>=0.12.3
